Refactor HTTPClient and Response
This change replaces the existing HTTPClient and Response models with a
newer version that is more complete in functoin and testing. Instead of
creating a self-dependent import from my own repo `http_overeasy`, the
library is being added directly in to this repo.

Correct dev-install and update phony

deps update

WIP 2